Adds multiple-user functionality as well as small style changes
I added the ability to have multiple users interact with the MIDCA agent
at the same time by creating special simulate and interpret modules
which connect to a socket server. The users are "remote" and can access
the MIDCA simulation only through the client. This limits them to seeing
what is given to the client and giving goals to the Agent. This is all
done through simple TCP sockets.

I also made it much cleaner to see the MIDCA agent move by clearing the
screen before every simulate phase.