Adds coins, changes goal choosing process
I added goal assessment code in the intend phase, which chooses a goal
based on a heuristic analyzing how fast the goal can be completed. the
goal which is expected to be completed the fastest is chosen as the goal
to act on.

I also added coin objects, though nothing has been done with them yet.