Adds an interactive dungeon creator and MIDCA description language.
I added some code to allow a user to interactively build and then save a dungeon, which is
much easier than designing one by hand. I also changed the way the code stores and reads
the dungeon object. THe biggest change though is that a Dungeon object can now spit out
its representation as a MIDCA state file in the planning language, and I reinstated a
domain description file.