Implements damage and begins adding bomb demo code
I added the ability for the agent to take damage from fire and traps, but I haven't yet altered pathfinding or planning code to handle that. So the
agent will happily run itself into fires and break down. I also added NPCs and the ability for the agent to bomb an area, killing NPCs. The 'killed' goal is what instructs the agent to do this. However, I haven't
implemeneted the 'killed' goal in terms of the user giving it to the agent.