Adds full rebellion with explanation
This major update implements a full rebellion demo domain, based on the Dungeon
domain. The agent is a drone and the operator can instruct it to bomb enemies
in order to eliminate them. The drone will not allow itself to bomb civilians,
so if it sees any civilians, it will rebel against the kill order. It is able
to suggest alternative enemies to be killed.

Communication between the world simulator, the agents and the users has also
been upgraded, so that Bilateral communications are possible between the
operators and the agents.