Adds automatic operator support for all sizes of maps
I fixed a bug that was causing data transmission to be truncated as a result
of large map sizes. Previously agents and operators received the pickled
version of objects they sought by expecting a fixed number of bytes. However,
the size of a map's or agent's pickle string changes, and increases as the
program goes on. So I changed the way clients recv data to gather every
message until it finds a terminal character ('\254').