Add variable compliance to agents
I added the ability to change the compliance of the agents with regards to the
operator's rejection of a rebellion. If the agent rebels against goal G and
the operator rejects this rebellion, then the agnet marks G as a mandatory
goal, which it will not rebel against again.

I also added some color to the display, so that agents are colored blue,
enemies are colored red, and civilians are colored green.